+++ src/wai/annotations/roi/constants.py
@@ -3,3 +3,6 @@ Module for constants relating to the ROI CSV format.
 """
 # The default extension for ROI CSV files
 DEFAULT_EXTENSION = "-rois.csv"
+
+# The symbol indicating a line in the ROI file is a comment
+COMMENT_SYMBOL = "#"